Add coerceTypes option to OpenAPIBackend and OpenAPIValidator
This commit introduces a new 'coerceTypes' option in both the OpenAPIBackend and OpenAPIValidator classes. When enabled, it allows for type coercion of request path and query parameters, enhancing the flexibility of request handling. The default value for this option is set to false. Additionally, the ValidationResult interface has been updated to include a 'coerced' property to store the coerced values.